requirements.txt
Flask==2.3.3
numpy==1.24.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
gunicorn==21.2.0
neo4j==5.15.0
python-dotenv==1.0.0
Flask-WTF==1.1.1
WTForms==3.0.1
pandas==2.0.3
